updated capabilities2_event, WIP capabilities2_runner and server
