+++ .devcontainer/Dockerfile
@@ -27,4 +27,6 @@ RUN rosdep update
 RUN echo "source /opt/ros/${ROS_DISTRO}/setup.bash" >> ~/.bashrc
 
 # make sure folders exist
-RUN mkdir -p ~/colcon_ws/src
+WORKDIR /home/ubuntu/colcon_ws/src
+
+RUN git clone https://github.com/CollaborativeRoboticsLab/capabilities2_test_suite.git
