+++ capabilities2_msgs/CMakeLists.txt
@@ -27,7 +27,7 @@ set(msg_files
   "msg/NaturalCapability.msg"
   "msg/Remapping.msg"
   "msg/RunningCapability.msg"
-  "msg/CapabilityOption.msg"
+  "msg/CapabilityParameter.msg"
 )
 
 set(srv_files
