+++ capabilities2_msgs/msg/CapabilityParameter.msg
