+++ capabilities2_runner/include/capabilities2_runner/notrigger_runner.hpp
@@ -16,7 +16,7 @@ class NoTriggerRunner : public RunnerBase
 {
 public:
   // throw on trigger function
-  void trigger(const capabilities2::CapabilityOptions& parameters, const std::string& bond_id) override
+  void trigger(const capabilities2::CapabilityParameters& parameters, const std::string& bond_id) override
   {
     // emit failed event
     emit_failed(bond_id, "cannot trigger this is a no-trigger runner");
