+++ capabilities2_runner/src/system/get_capability_specs_runner.cpp
@@ -1,6 +1,6 @@
 #include <pluginlib/class_list_macros.hpp>
 #include <capabilities2_runner/runner_base.hpp>
-#include <capabilities2_runner/get_capability_specs_runner.hpp>
+#include <capabilities2_runner/system/get_capability_specs_runner.hpp>
 
 // register runner plugins
 PLUGINLIB_EXPORT_CLASS(capabilities2_runner::GetCapabilitySpecsRunner, capabilities2_runner::RunnerBase);
