+++ capabilities2_runner/src/system/multiplex_all_runner.cpp
@@ -1,6 +1,6 @@
 #include <pluginlib/class_list_macros.hpp>
 #include <capabilities2_runner/runner_base.hpp>
-#include <capabilities2_runner/input_multiplex_all_runner.hpp>
+#include <capabilities2_runner/system/input_multiplex_all_runner.hpp>
 
 // register runner plugins
 PLUGINLIB_EXPORT_CLASS(capabilities2_runner::InputMultiplexAllRunner, capabilities2_runner::RunnerBase);
