--- capabilities2_test_suite/action/Fibonacci.action
@@ -1,5 +1,0 @@
-int32 order
----
-int32[] sequence
----
-int32[] partial_sequence
