--- capabilities2_test_suite/package.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>capabilities2_test_suite</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="kalanaratnayake95@gmail.com">ubuntu</maintainer>
-  <license>MIT</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>rclcpp</depend>
-  <depend>pluginlib</depend>
-  <depend>capabilities2_runner</depend>
-  <depend>capabilities2_msgs</depend>
-  <depend>builtin_interfaces</depend>
-  <depend>rosidl_default_generators</depend>
-  <depend>rosidl_default_runtime</depend>
-
-   <member_of_group>rosidl_interface_packages</member_of_group>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-
-    <!-- TestActionRunner -->
-    <capability_interface>
-      interfaces/TestActionRunner.yaml
-    </capability_interface>
-    
-    <capability_provider>
-      providers/TestActionRunner.yaml
-    </capability_provider>
-
-
-  </export>
-</package>
