--- capabilities2_test_suite/providers/TestActionRunner.yaml
@@ -1,9 +1,0 @@
-# the empty provider
-%YAML 1.1
----
-name: TestActionRunner
-spec_type: provider
-spec_version: 1.1
-description: "The capability provider for the test action runner interface"
-implements: capabilities2_test_suite/TestActionRunner
-runner: capabilities2_runner::TestActionRunner
