--- capabilities2_test_suite/src/test_runners.cpp
@@ -1,5 +1,0 @@
-#include <pluginlib/class_list_macros.hpp>
-#include <capabilities2_runner/runner_base.hpp>
-#include <capabilities2_test_suite/test_runners/test_action_runner.hpp>
-
-PLUGINLIB_EXPORT_CLASS(capabilities2_runner::TestActionRunner, capabilities2_runner::RunnerBase);
